EDIT Javadoc for Play and Open methods, and fixed error in AudioMixer
Documentation now explains that Play and Start can be invoked when the
AudioCue has not been opened. This allows one to call the
AudioCuePlayer.readTrack method and obtain PCM data without writing it
to the system out. Also, an error in AudioMixer Javadoc was fixed,
changing the documentation on the default Buffer size to the correct
8192 bytes.